requirements.txt
django==2.2.1
Pillow
